JCSP-419:DC,Cloud: project issue navigator. Implemented PR comments
