+++ custom-vu/src/main/kotlin/jces1209/vu/JiraDcScenario.kt
@@ -8,13 +8,13 @@ import com.atlassian.performance.tools.jiraactions.api.measure.ActionMeter
 import com.atlassian.performance.tools.jiraactions.api.memories.UserMemory
 import com.atlassian.performance.tools.jiraactions.api.scenario.JiraCoreScenario
 import com.atlassian.performance.tools.jiraactions.api.scenario.Scenario
-import jces1209.vu.action.*
+import jces1209.vu.action.CreateAnIssue
+import jces1209.vu.action.ProjectIssueNavigatorAction
+import jces1209.vu.action.SearchServerFilter
 import jces1209.vu.page.DcIssuePage
-import jces1209.vu.page.JiraTips
 import jces1209.vu.page.boards.browse.dc.DcBrowseBoardsPage
-import jces1209.vu.page.boards.view.BoardPage
 import jces1209.vu.page.filters.ServerFiltersPage
-import jces1209.vu.page.issue.DcProjectIssueNavigatorPage
+import jces1209.vu.page.project.DcProjectIssueNavigatorPage
 import org.openqa.selenium.By
 import org.openqa.selenium.TakesScreenshot
 
@@ -59,6 +59,13 @@ class JiraDcScenario : Scenario {
                 jira = jira,
                 meter = meter,
                 projectMemory = similarities.projectMemory
+            ),
+            projectIssueNavigatorAction = ProjectIssueNavigatorAction(
+                jira = jira,
+                meter = meter,
+                projectMemory = similarities.projectMemory,
+                projectIssueNavigatorPage = DcProjectIssueNavigatorPage(jira.driver),
+                numberOfProject= 3
             )
         )
     }
