+++ custom-vu/src/main/kotlin/jces1209/vu/action/LogInWithCookie.kt
@@ -8,8 +8,7 @@ import jces1209.vu.page.JiraCloudWelcome
 import org.openqa.selenium.Cookie
 import java.time.ZonedDateTime
 import java.time.format.DateTimeFormatter
-import java.util.Date
-import java.util.Properties
+import java.util.*
 
 /**
  * Logs in via Cloud Session cookie. Reads from `cloud-session.properties` classpath resource.
