+++ custom-vu/src/main/kotlin/jces1209/vu/action/ProjectIssueNavigatorAction.kt
@@ -5,7 +5,7 @@ import com.atlassian.performance.tools.jiraactions.api.WebJira
 import com.atlassian.performance.tools.jiraactions.api.action.Action
 import com.atlassian.performance.tools.jiraactions.api.measure.ActionMeter
 import com.atlassian.performance.tools.jiraactions.api.memories.ProjectMemory
-import jces1209.vu.page.issue.AbstractProjectIssueNavigatorPage
+import jces1209.vu.page.project.ProjectIssueNavigatorPage
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 
@@ -13,8 +13,8 @@ class ProjectIssueNavigatorAction(
     private val jira: WebJira,
     private val meter: ActionMeter,
     private val projectMemory: ProjectMemory,
-    private val projectIssueNavigatorPage: AbstractProjectIssueNavigatorPage,
-    private val numberOfProject: String
+    private val projectIssueNavigatorPage: ProjectIssueNavigatorPage,
+    private val numberOfProject: Int
 
 ) : Action {
     private val logger: Logger = LogManager.getLogger(this::class.java)
