+++ custom-vu/src/main/kotlin/jces1209/vu/page/CloudIssueLinking.kt
@@ -6,7 +6,6 @@ import org.openqa.selenium.Keys
 import org.openqa.selenium.WebDriver
 import org.openqa.selenium.interactions.Actions
 import org.openqa.selenium.support.ui.ExpectedCondition
-import org.openqa.selenium.support.ui.ExpectedConditions
 import org.openqa.selenium.support.ui.ExpectedConditions.*
 
 class CloudIssueLinking(
