+++ custom-vu/src/main/kotlin/jces1209/vu/page/FalliblePage.kt
@@ -1,12 +1,11 @@
 package jces1209.vu.page
 
+import com.atlassian.performance.tools.jiraactions.api.page.wait
 import org.openqa.selenium.By
 import org.openqa.selenium.WebDriver
+import org.openqa.selenium.support.ui.ExpectedCondition
 import org.openqa.selenium.support.ui.ExpectedConditions.or
 import org.openqa.selenium.support.ui.ExpectedConditions.visibilityOfElementLocated
-import com.atlassian.performance.tools.jiraactions.api.page.wait
-import org.openqa.selenium.support.ui.ExpectedCondition
-import org.openqa.selenium.support.ui.ExpectedConditions
 import java.time.Duration
 
 class FalliblePage private constructor(
