+++ custom-vu/src/main/kotlin/jces1209/vu/page/JiraCloudWelcome.kt
@@ -3,9 +3,7 @@ package jces1209.vu.page
 import jces1209.vu.wait
 import org.openqa.selenium.By
 import org.openqa.selenium.WebDriver
-import org.openqa.selenium.support.ui.ExpectedConditions.elementToBeClickable
-import org.openqa.selenium.support.ui.ExpectedConditions.or
-import org.openqa.selenium.support.ui.ExpectedConditions.presenceOfElementLocated
+import org.openqa.selenium.support.ui.ExpectedConditions.*
 
 class JiraCloudWelcome(
     private val driver: WebDriver
