--- custom-vu/src/main/kotlin/jces1209/vu/page/issue/AbstractProjectIssueNavigatorPage.kt
@@ -1,5 +1,0 @@
-package jces1209.vu.page.issue
-
-interface AbstractProjectIssueNavigatorPage {
-    fun openProjectByIndex(numberOfProject: String)
-}
