+++ src/test/kotlin/JiraPerformanceComparisonIT.kt
@@ -12,14 +12,12 @@ import com.atlassian.performance.tools.virtualusers.api.config.VirtualUserTarget
 import com.atlassian.performance.tools.workspace.api.RootWorkspace
 import com.atlassian.performance.tools.workspace.api.TestWorkspace
 import jces1209.BenchmarkQuality
-import jces1209.QuickAndDirty
 import jces1209.SlowAndMeaningful
 import jces1209.log.LogConfigurationFactory
 import jces1209.vu.JiraCloudScenario
 import jces1209.vu.JiraDcScenario
 import org.apache.logging.log4j.core.config.ConfigurationFactory
 import org.junit.Test
-import java.io.File
 import java.nio.file.Paths
 import java.time.Duration
 import java.util.concurrent.CompletableFuture
